+++ processor/src/test/fixtures/input/com/example/multiround/MyFragment.java
@@ -1,0 +1,8 @@
+package com.example.multiround;
+
+import sting.Fragment;
+
+@Fragment( includes = MyGeneratedFragment.class )
+public interface MyFragment
+{
+}