+++ processor/src/test/fixtures/input/com/example/multiround/MyGeneratedFragment.java
@@ -1,0 +1,10 @@
+package com.example.multiround;
+
+@sting.Fragment
+public interface MyGeneratedFragment
+{
+  default Object provideObject()
+  {
+    return new Object();
+  }
+}