+++ processor/src/test/fixtures/input/com/example/multiround/MyInjector.java
@@ -1,0 +1,9 @@
+package com.example.multiround;
+
+import sting.Injector;
+
+@Injector( includes = MyFragment.class )
+interface MyInjector
+{
+  Object getObject();
+}